feat: add WIP modem interface for downloader
Note: I haven't done much stuff with this. Also, as the download phone
number (18004953555) is hard-coded into the system, simply having
access to two landlines is unfortunately not enough to experiment with
this. You'll need some way of giving it a fake phone line to connect
to, such as a telephone line simulator with programmable numbers. I use
a Teltone TLS-4 for this purpose.