Revert "feat: change FW version # in patch"
Download server was hanging up on me and I'm pretty sure this was why.
Could be a good thing if you don't want your programming modified w/o
your permission, but then the proper solution would be to block it on
your end, not trick them into hanging up. XD

This reverts commit 465ea03488f1bae5249a0cfa13465fd606322cf7.